Implemented the following:
Added the ability to upload files from a computer, as well as download from the server.
The AppDisk file has been changed: all components have been separated into separate files.
Added the ability to change the type of files: tabular and grid.
Added the ability to delete a file.
It was not possible to add changes to the git repository, since the Internet ended :(